+++ pac4j-core/src/main/java/org/pac4j/core/util/JavaSerializationHelper.java
@@ -4,10 +4,9 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.io.*;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.Base64;
-import java.util.List;
+import java.util.*;
+import java.util.function.Function;
+import java.util.stream.Collectors;
 
 /**
  * Helper for Java serialization.
@@ -19,12 +18,15 @@ public class JavaSerializationHelper {
 
     private static final Logger logger = LoggerFactory.getLogger(JavaSerializationHelper.class);
 
-    private List<String> trustedPackages;
+    private Set<String> trustedPackages;
+
+    private Set<Class<?>> trustedClasses;
 
     public JavaSerializationHelper() {
-        trustedPackages = new ArrayList<>();
+        trustedPackages = new HashSet<>();
         trustedPackages.addAll(Arrays.asList("java.", "javax.", "org.pac4j.", "com.github.scribejava.", "org.opensaml.", "com.nimbusds.",
             "[Lcom.nimbusds.", "org.joda.", "net.minidev.json.", "org.bson.types."));
+        trustedClasses = new HashSet<>();
     }
 
     /**
@@ -57,25 +59,25 @@ public class JavaSerializationHelper {
     }
 
     /**
-     * Unserialize a base64 String into a Java object.
+     * Deserialize a base64 String into a Java object.
      *
      * @param base64 the serialized object as a base64 String
-     * @return the unserialized Java object
+     * @return the deserialized Java object
      */
-    public Serializable unserializeFromBase64(final String base64) {
-        return unserializeFromBytes(Base64.getDecoder().decode(base64));
+    public Serializable deserializeFromBase64(final String base64) {
+        return deserializeFromBytes(Base64.getDecoder().decode(base64));
     }
 
     /**
-     * Unserialize a bytes array into a Java object.
+     * Deserialize a bytes array into a Java object.
      *
      * @param bytes the serialized object as a bytes array
-     * @return the unserialized Java object
+     * @return the deserialized Java object
      */
-    public Serializable unserializeFromBytes(final byte[] bytes) {
+    public Serializable deserializeFromBytes(final byte[] bytes) {
         Serializable o = null;
         try (final ByteArrayInputStream bais = new ByteArrayInputStream(bytes);
-            final ObjectInputStream ois = new RestrictedObjectInputStream(bais, this.trustedPackages)) {
+            final ObjectInputStream ois = new RestrictedObjectInputStream(bais, this.trustedPackages, this.trustedClasses)) {
             o = (Serializable) ois.readObject();
         } catch (final IOException | ClassNotFoundException e) {
             logger.warn("cannot Java deserialize object", e);
@@ -83,17 +85,51 @@ public class JavaSerializationHelper {
         return o;
     }
 
-    public List<String> getTrustedPackages() {
-        return trustedPackages;
+    /**
+     * Returns an immutable set of tusted packages.
+     *
+     * @return the trusted packages
+     */
+    public Set<String> getTrustedPackages() {
+        return Collections.unmodifiableSet(trustedPackages);
     }
 
-    public void setTrustedPackages(final List<String> trustedPackages) {
-        this.trustedPackages = trustedPackages;
+    /**
+     * Returns an immutable set of trusted classes.
+     *
+     * @return the trusted classes
+     */
+    public Set<Class<?>> getTrustedClasses() {
+        return Collections.unmodifiableSet(trustedClasses);
+    }
+
+    public void addTrustedPackages(final Collection<String> trustedPackages) {
+        this.trustedPackages.addAll(trustedPackages);
+    }
+
+    public void addTrustedPackage(final String trustedPackage) {
+        this.trustedPackages.add(trustedPackage);
+    }
+
+    public void clearTrustedPackages() {
+        this.trustedPackages.clear();
+    }
+
+    public void addTrustedClasses(final Collection<Class<?>> trustedClasses) {
+        this.trustedClasses.addAll(trustedClasses);
+    }
+
+    public void addTrustedClass(final Class<?> trustedClass) {
+        this.trustedClasses.add(trustedClass);
+    }
+
+    public void clearTrustedClasses() {
+        this.trustedClasses.clear();
     }
 
     @Override
     public String toString() {
-        return CommonHelper.toNiceString(this.getClass(), "trustedPackages", this.trustedPackages);
+        return CommonHelper.toNiceString(this.getClass(), "trustedPackages", this.trustedPackages, "trustedClasses", this.trustedClasses);
     }
 
     /**
@@ -101,27 +137,32 @@ public class JavaSerializationHelper {
      */
     private static class RestrictedObjectInputStream extends ObjectInputStream {
 
-        private final List<String> trustedPackages;
+        private final Set<String> trustedPackages;
 
-        private RestrictedObjectInputStream(final InputStream in, final List<String> trustedPackages) throws IOException {
+        private final Map<String, Class<?>> trustedClasses; // className -> Class
+
+        private RestrictedObjectInputStream(final InputStream in, final Set<String> trustedPackages,
+                                            final Set<Class<?>> trustedClasses) throws IOException {
             super(in);
-            CommonHelper.assertNotNull("trustedPackages", trustedPackages);
             this.trustedPackages = trustedPackages;
+            this.trustedClasses = trustedClasses.stream().collect(Collectors.toMap(Class::getName, Function.identity()));
         }
 
         @Override
         protected Class<?> resolveClass(final ObjectStreamClass desc) throws IOException, ClassNotFoundException {
             final String qualifiedClassName = desc.getName();
-            for (final String p : trustedPackages) {
-                if (qualifiedClassName.startsWith(p)) {
-                    return super.resolveClass(desc);
-                }
+            final Class<?> clazz = trustedClasses.get(qualifiedClassName);
+            if (Objects.nonNull(clazz)) {
+                return clazz;
+            } else if (trustedPackages.stream().anyMatch(qualifiedClassName::startsWith)) {
+                return super.resolveClass(desc);
+            } else {
+                throw new ClassNotFoundException("Wont resolve untrusted class: " + qualifiedClassName);
             }
-            throw new ClassNotFoundException("Wont resolve untrusted class: " + qualifiedClassName);
         }
 
         @Override
-        protected Class<?> resolveProxyClass(String[] interfaces) throws IOException, ClassNotFoundException {
+        protected Class<?> resolveProxyClass(String[] interfaces) throws ClassNotFoundException {
             throw new ClassNotFoundException("Wont resolve proxy classes at all.");
         }
     }